# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 智岗罗盘 API
  version: 1.0.0
  description: |
    多源异构驱动的岗位能力动态演化与人岗匹配系统 API 契约。
    本文件是前后端并行开发的单一事实源（AGENTS.md 铁律一）。
    前端通过 `openapi-typescript` 生成类型，不得自行定义与后端不一致的类型。
    认证方案：Bearer 双 Token（access 30min / refresh 7d）；refresh_token 同时
    写入 httpOnly Cookie（httponly + samesite=lax），供页面刷新后静默恢复会话。
    标注规则：全部端点已实现（M1-M4 全量交付），契约不含里程碑占位标注。
  contact:
    name: 智岗罗盘团队
    url: https://github.com/zhigang-compass
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.zhigang.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/admin/crawl/trigger:
    post:
      tags: [admin]
      summary: 触发爬取任务（平台白名单校验 + ARQ 入队）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrawlTriggerRequest'
      responses:
        '202':
          description: 任务已触发（data 含 task_id/status=pending）
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/CrawlTriggerResult'
                  trace_id: { type: string }
...
  /api/v1/admin/crawl/history:
    get:
      tags: [admin]
      summary: 爬取历史（task_status 中 crawl 任务，倒序分页）
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 爬取历史列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  msg: { type: string }
                  data:
                    $ref: '#/components/schemas/CrawlHistoryData'
                  trace_id: { type: string }
  /api/v1/admin/crawl/task/{task_id}/stream:
    get:
      tags: [admin]
      summary: "爬虫实时日志 SSE（event: log/progress/done/error）"
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: SSE 流
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
...
  schemas:
...
    CrawlTriggerRequest:
      type: object
      required: [platform]
      properties:
        platform: { type: string, minLength: 1, maxLength: 50, description: "平台 ID：boss/zhilian/monster/indeed/glassdoor/linkedin/maimai/github/stackoverflow/arxiv/icourse163/coursera/edx" }
        keyword: { type: string, maxLength: 200, description: "关键词（可选，留空则采集平台热度/最新内容，08-16 起爬虫不再内置默认关键词）" }
        city: { type: string, maxLength: 50, description: "城市（可选，海外源默认英文城市）" }
...
    CrawlHistoryItem:
      type: object
      description: 爬取任务历史项（GET /admin/crawl/history）
      required: [id, platform, platform_name, keyword, status, items, error]
      properties:
        id: { type: string }
        platform: { type: string, description: spider 名 }
        platform_name: { type: string, description: 平台中文名 }
        keyword: { type: string }
        status: { type: string, enum: [pending, running, success, failed] }
        items: { type: integer, description: 入库条数 }
        error: { type: string }
        created_at: { type: string, nullable: true }
    CrawlHistoryData:
      type: object
      description: GET /admin/crawl/history 响应 data（分页）
      required: [items, total, page, size]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CrawlHistoryItem' }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
    CrawlTriggerResult:
      type: object
      description: POST /admin/crawl/trigger 响应 data（ARQ 入队结果）
      required: [task_id, platform, status]
      properties:
        task_id: { type: string }
        platform: { type: string }
        status: { type: string }
...